+++ frontend/src/content/changelog.md
@@ -10,6 +10,7 @@ description: 'A list of all changes made to the game'
 
 - follow Ranker should now try to center your Ranker
 - the game should now be iframe-able to allow for galaxy.click to embed it
+- FairWiki now opens in a new tab
 
 <br><br><br><br><br><br><br>